Replace tweaks with features (for now)
